init AutoMovementCommand and DriveTrain migration
sorry for large commit!

I changed DriveTrain to an abstract class and removed all of the methods that were only being implemented in DriveTrainC from it. This structure better reflects the relationship between DriveTrain, DriveTrainC, and DriveTrainP.

Because of the above changes, every command which now uses DriveTrainC only methods had to be updated to require a DriveTrainC specifically (previously every command only required a DriveTrain). This structure better reflects the actual dependencies of our classes.